circular buffers
Former-commit-id: 397634452113cdf5af80100b805d0242b3f61267 [formerly 7fdb37abb4ddf6de4293fcab05ac28d45f48343b] [formerly 5be7539bc64ea5fff98713d3d1e88289bc4884ae [formerly 1f90fdd24acf087d4f2ed8609bd652f6af85b656]] [formerly 4047b764201d7ab5b1be4a46fc18bbc1476895e4 [formerly a87858f9fbc8cb72e9e612a6413671dae4904ece]]
Former-commit-id: 4047b764201d7ab5b1be4a46fc18bbc1476895e4
Former-commit-id: 8311ae9d456b09b4a22d6f262e70dcd2511735b6 [formerly 2989fd4a38a7f7b0cf40dd7518745fa44acf6b1d]
Former-commit-id: e71ec3ae15e354894a84dfc9f5d1625d82ba96be

+++ templating/templates/utilityFunctions.c
@@ -3,7 +3,6 @@
 #include <fcntl.h>
 #include <unistd.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include "constants.h"
 
@@ -83,7 +82,7 @@ void set_sighandler(int signum, void *psh, sigset_t *block_mask) {
 * If the file descriptor needs to be ftruncated (i.e., this is the first process opening)
 * this shared memory, then make sure the O_CREAT flag is set in shm_flags
 */
-void open_shared_mem(uint8_t **ppmem, const char *pName, int numBytes, int shm_flags, int mmap_flags) {
+void open_shared_mem(uint8_t **ppmem, const char *pName, size_t numBytes, int shm_flags, int mmap_flags) {
   int fd = shm_open(pName, shm_flags, 0600);
   if(fd == -1) {
     die("shm_open failed\n");
